+++ src/main/java/fr/raksrinana/fallingtree/config/CommonConfig.java
@@ -19,6 +19,7 @@ public class CommonConfig{
 	public final ForgeConfigSpec.BooleanValue ignoreDurabilityLoss;
 	public final ForgeConfigSpec.BooleanValue preserveTools;
 	public final ForgeConfigSpec.BooleanValue reverseSneaking;
+	public final ForgeConfigSpec.BooleanValue breakLeaves;
 	public final ForgeConfigSpec.IntValue maxTreeSize;
 	
 	
@@ -32,6 +33,7 @@ public class CommonConfig{
 		maxTreeSize = builder.comment("The maximum size of a tree. If there's more logs than this value the tree won't be cut.").defineInRange("max_log_count", 100, 1, Integer.MAX_VALUE);
 		preserveTools = builder.comment("When set to true, when a tree is broken and the tool is about to break we will just break one block and not the whole tree.").define("preserve_tools", false);
 		reverseSneaking = builder.comment("When set to true, a tree will only be chopped down if the player is sneaking").define("reverse_sneaking", false);
+		breakLeaves = builder.comment("When set to true, leaves will be broken instantly").define("break_leaves", false);
 	}
 	
 	public Stream<Block> getWhitelistedLogs(){