+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## [1.15.1-2.0.1] - 2019-12-20
+- Add an option in the config to instantly break leaves (defaults to false).
+
 ## [1.15.1-2.0.0] - 2019-12-17
 - Port to 1.15.1